Replace nearly all literal codes with symbols
To provide better mnemonics and searching for the message in question,
and cross referencing with the "Message Flow" chapter of the
PostgreSQL manual.  The literal value of the message type is retained
at the end of its corresponding constant name.

This is not a completely mechanical replacement because some message
type characters are reused in various protocol states, and in those
cases the correct code name is to be used.  These names and order of
messages in "messages.go" are copied directly from the PostgreSQL 9.1
manual.

Signed-off-by: Dan Farina <drfarina@acm.org>