Confirm that Postgres does not allow bad cstrings for param values
Looks like it is checked in postgres.c:exec_bind_message if the format
of parameters is textual...which seems like all of them right now.

It also looks like for binary mode that Postgres may protect against
this, by ensuring that a type receive function eats the entire
parameter value buffer.  A poorly written Postgres recv (binary
decoding) function could open the user to nul byte injections, but I'm
not sure if there's much to be done about that.

Signed-off-by: Daniel Farina <daniel@heroku.com>