Adding Hstore support package
Includes two Hstore types, for the common non-null values use case
(Hstore) and when nulls are necessary (HstoreWithNulls).

A map alias is not used to keep things obvious and allow for NULL
columns without adding more types and simplify Scan() parameters.

A test suite with a variety of values is also included.