Support the "escape" bytea format when receiving data
This allows pq to operate correctly in both output formats, which
should help users who are still on PG 9.0 or have resorted to
setting bytea_output to "escape" for other reasons.

Even after this commit, bytea data is always sent in the "hex"
format.  This is not a problem for PostgreSQL versions 9.0 and later
as they are guaranteed to accept either format.  Older versions will
continue to experience problems, but it's not clear whether that
problem is worth fixing or not.