Don't allow zero bytes in the result of QuoteIdentifier
Because a protocol-level String is null-terminated, there would have
been a possibility of a specially crafted input string manipulating the
protocol packet.  However, the worst thing I can see is that the
attacker might be able to specify the input parameter types of a query
in a Parse message, which doesn't seem too problematic.