Display a nicer error when a query is run with more than 2^16-1 params
Because the number of parameters in the wire protocol is an uint16,
there's no way to execute a prepared statement with more than 2^16-1
parameters.  Display a nicer error message in that case.

The error will still be a bit obscure if the user explicitly prepared
the statement, though, since database/sql will be the one complaining in
that case.

I also deliberately did not commit a test case for this, since running
the test I wrote took more than 20 seconds on my laptop, which isn't
justifiable for a corner case like this.  I've left my test case into
issue #295.

Per complaint from Brian Fallik in the same issue.