Disable SSL when connecting over a UNIX domain socket
Previously if sslmode was not explicitly disabled, trying to connect
over a UNIX domain socket would result in a "pq: SSL is not enabled on
the server" error.  This is both silly and incompatible with libpq.  Fix
by setting "sslmode" to "disable" automatically whenever a connection is
made over a UNIX socket.