ListenerConn: Fix deadlock when connection is lost while a query is queued
Because the two goroutines acquire the two exclusive locks in different
orders, they would deadlock with each other for no reason.  Fix by
swapping the order of acquisition in acquireSenderLock.  If l.err is set
nobody should be holding senderLock for a long time anyway, so undesired
practical effects of this change should be very minimal.