Implement driver option binary_parameters
If set, all parameters of type []byte will be sent over as binary data,
which allows the driver to avoid the second round-trip to the server
when not using a prepared statement proper.  This also allows pgbouncer
to be used with parameterized queries, which is impossible if
binary_parameters has not been enabled.

This patch is still missing documentation, but the code has been sitting
on my hard drive for long enough.  I'll commit the docs in a followup
commit.

Marko Tiikkaja, with some help from Chris Gilling