Make copyin.Close() idempotent
This change is causing breakage on tip:
https://github.com/golang/go/issues/12798

I think it's a reasonable upstream change.

Looking at the commit that added the code that is failing on our side
the lack of error propagation is called out and is the reason we're now
failing.
https://github.com/lib/pq/commit/72f84510ec4899677ec603abbaec5ba17a438a8c

This change attempts to preserve old behavior while supporting the fact
that errors are now propagated properly by Stmt.Close().

Open to better ways to do this though.