Don't auto-close on final CopyIn Exec() and don't return invalid Result
The docs suggest a final empty `stmt.Exec()` at the end of a CopyIn operation
however this is no longer strictly required (as discussed in #444).

This resolves #444 by not auto-closing on an empty `stmt.Exec()` and by fixing
the docs. Code that performs an empty `stmt.Exec()` will continue to work
properly as long as there is an explicit `stmt.Close()` which is already
required by the `database/sql` API anyway.

It also resolves #626 in case any code actually uses the `sql.Result` returned
by an empty `stmt.Exec()`.