fix unsafe rows.Close() race
According to https://godoc.org/database/sql/driver#Rows, the Close
operation should not affect the values returned by Scan. In normal flow,
this is not an issue, but if Rows is closed by a cancelled context or
another routine, it can be.

This changes the scratch to be a slice, and changes its memory location
when rows are closed before they are done, so that any scanning will not
be affected.