Add support for database/sql Connector and OpenDB
What
===
Add support for `database/sql/driver` `Connector` and `database/sql`
`OpenDB` that were added in go1.10.

Why
===
To allow users to open a connection without using the registered drivers
interface, and instead explicitly getting a `Connector` from the this
driver and passing that connector to `database/sql` `OpenDB`.

Notes
===
- The new types and functions are only built for go1.10 or newer because
the `database/sql/driver` `Connector` interface and `database/sql`
`OpenDB` function are only available in go1.10 or later.
- The `NewConnector` has support for context, but the context isn't used
at this point. Use of it could be added if desired, but that could occur
separately to this change.