Add `QuoteLiteral` function.
There already exists a `QuoteIdentifier` function that helps to escape
identifiers that you need to dynamically generate. `QuoteLiteral` does the
same, but for literals, which is helpful for DDL statements that utilize literals,
as you cannot pass parameters into said statements.

This is modeled after the way libpq implements `PQescapeStringInternal` and
can be referenced here:

https://git.postgresql.org/gitweb/?p=postgresql.git;a=blob;f=src/interfaces/libpq/fe-exec.c;hb=HEAD#l3336