only retry sending on network errors if nothing was transferred
It is possible that a query string contains multiple SQL statements.
Depending on the server implementation, it might be possible that the
server executes the SQL statements as soon as it receives a
termination character.
If a net.OpError happened afterwards we can not send the whole
query string again to the server because the already executed statements
would be executed again.

Add an additional check in mustSendRetryable() to only return a
netErrorNoWrite error if the number of bytes that were send out is 0.