Fix sslinline for connections after the first one.
The sslinline functionality would delete the required keys after
upgrading the first connection.
This commit adds the sslinine key to
the isDriverSetting so that it is not sent to
Posgres which would cause a crash.
It also removes the delete of the keys necessary for sslinline to allow
SSL upgrades for connections after the first one.