Clarify maintenance mode
Looking through recent issues, this seems to be roughly the current situation. Making it clearer here, to help users evaluating this lightweight library against the heavier-weight pgx. But I confess, I might be guessing totally wrong.

Whether or not this PR content is right, I just want to say thanks to the maintainers of this library. We're using it at Camus Energy to help grid operators decarbonize, a.k.a. fighting climate change. We're really happy this library exists, because it serves our needs better than pgx. Really, thanks 😃