Fix flaky TestConnPrepareContext
TestConnPrepareContext checks it receives an context.DeadlineExceeded. However,
the context isn't necessarily always expired. This causes most of the test runs
to fail for me, only occasionally succeeding. This change ensures the context
used in the test is actually context.DeadlineExceeded. The negative duration
causes the context package to return an already canceled context.