Don't reset extra_float_digits
In PostgreSQL ≤11 the extra_float_digits parameter controlled how many
digits were used for floating point numbers. It was set here to ensure
people didn't accidentally needlessly lose precision (in 78a464e it was
set to 3. In 5941153 it was changed to 2 as people reported problems on
PostgreSQL ≤8.4, see #199).

Since PostgreSQL 12 it will always use maximum precision unless it's set
to ≤0. We only support PostgreSQL ≥14 now, so we can just remove it
here.

Ref: https://www.postgresql.org/docs/current/datatype-numeric.html#DATATYPE-FLOAT

Fixes #475
Fixes #1150
Fixes #1146